refactor(spf): home non-behavior playback modules in playback/primitives
failoverFetch is signal-aware but has no behavior lifecycle (no
effect/computed/subscribe), so sitting in playback/behaviors/ misread it
as a behavior. Introduce playback/primitives/ for core-aware playback
building blocks that behaviors/actors/engines compose but that aren't
themselves behaviors, mirroring media/primitives/ one layer down. Move
failover-fetch and track-types (the other non-behavior) there, wire the
tsconfig project + vitest project, and document the category in
src/CLAUDE.md.
