+++ packages/spf/src/playback/primitives/tsconfig.json
@@ -4,16 +4,13 @@
     "composite": true,
     "lib": ["ES2022", "WebWorker"],
     "exactOptionalPropertyTypes": false,
-    "declarationDir": "../../../types/playback/behaviors"
+    "declarationDir": "../../../types/playback/primitives"
   },
   "references": [
     { "path": "../../../../utils" },
     { "path": "../../core" },
     { "path": "../../network" },
-    { "path": "../../media" },
-    { "path": "../primitives" },
-    { "path": "../actors" }
+    { "path": "../../media" }
   ],
-  "include": ["./*.ts", "./tests/**/*.ts"],
-  "exclude": ["./dom/**"]
+  "include": ["./*.ts", "./tests/**/*.ts"]
 }
