+++ tsconfig.json
@@ -23,6 +23,7 @@
     { "path": "packages/spf/src/network" },
     { "path": "packages/spf/src/media" },
     { "path": "packages/spf/src/media/dom" },
+    { "path": "packages/spf/src/playback/primitives" },
     { "path": "packages/spf/src/playback/behaviors" },
     { "path": "packages/spf/src/playback/behaviors/dom" },
     { "path": "packages/spf/src/playback/actors" },
